Fixed suggestions bug
Some suggestions requests might in some cases take longer than other
ones and if an older request takes longer than a newer one, then it
might overwrite the newer ones.

The suggestions now keep a time of request and it will only overwrite
older suggestions if the time of request is newer.